Moving JSON/YAML parsers to better understanding
